very huge advance

+++ Final_Proj/GameComponent/Chartlet.hpp
@@ -5,11 +5,11 @@ class Chartlet
 {
     public:
     char *aprnc;
-    long *aprnc_Xofst;
-    long *aprnc_Yofst;
+    int *aprnc_Xofst;
+    int *aprnc_Yofst;
     int length;
 
-    Chartlet(char* _aprnc, long* _aprnc_Xofst, long* _aprnc_Yofst, int _length)
+    Chartlet(char* _aprnc, int* _aprnc_Xofst, int* _aprnc_Yofst, int _length)
     {
         aprnc = _aprnc;
         aprnc_Xofst = _aprnc_Xofst;
@@ -20,9 +20,9 @@ class Chartlet
 
 extern char gnd_aprnc_1_element[];
 
-extern long gnd_aprnc_1_Xofst[];
+extern int gnd_aprnc_1_Xofst[];
 
-extern long gnd_aprnc_1_Yofst[];
+extern int gnd_aprnc_1_Yofst[];
 
 extern int gnd_aprnc_1_length;
 
@@ -30,9 +30,9 @@ extern int gnd_aprnc_1_length;
 
 extern char gnd_aprnc_2_element[];
 
-extern long gnd_aprnc_2_Xofst[];
+extern int gnd_aprnc_2_Xofst[];
 
-extern long gnd_aprnc_2_Yofst[];
+extern int gnd_aprnc_2_Yofst[];
 
 extern int gnd_aprnc_2_length;
 
@@ -40,9 +40,9 @@ extern int gnd_aprnc_2_length;
 
 extern char exc_aprnc_1_element[];
 
-extern long exc_aprnc_1_Xofst[];
+extern int exc_aprnc_1_Xofst[];
 
-extern long exc_aprnc_1_Yofst[];
+extern int exc_aprnc_1_Yofst[];
 
 extern int exc_aprnc_1_length;
 
@@ -50,9 +50,9 @@ extern int exc_aprnc_1_length;
 
 extern char exc_aprnc_2_element[];
 
-extern long exc_aprnc_2_Xofst[];
+extern int exc_aprnc_2_Xofst[];
 
-extern long exc_aprnc_2_Yofst[];
+extern int exc_aprnc_2_Yofst[];
 
 extern int exc_aprnc_2_length ;
 
@@ -60,9 +60,9 @@ extern int exc_aprnc_2_length ;
 
 extern char photon_aprnc_exist_element[];
 
-extern long photon_aprnc_exist_Xofst[];
+extern int photon_aprnc_exist_Xofst[];
 
-extern long photon_aprnc_exist_Yofst[];
+extern int photon_aprnc_exist_Yofst[];
 
 extern int photon_aprnc_exist_length;
 
@@ -70,9 +70,9 @@ extern int photon_aprnc_exist_length;
 
 extern char photon_aprnc_gone_element[];
 
-extern long photon_aprnc_gone_Xofst[];
+extern int photon_aprnc_gone_Xofst[];
 
-extern long photon_aprnc_gone_Yofst[];
+extern int photon_aprnc_gone_Yofst[];
 
 extern int photon_aprnc_gone_length;
